size_t to int

+++ set-matrix-zeroes.cpp
@@ -1,26 +1,26 @@
 class Solution {
 public:
     void setZeroes(vector<vector<int> > &matrix) {
-        const size_t m = matrix.size();
-        const size_t n = matrix[0].size();
+        const int m = matrix.size();
+        const int n = matrix[0].size();
         vector<bool> row(m, false);
         vector<bool> col(n, false);
 
-        for (size_t i = 0; i < m; ++i) {
-            for (size_t j = 0; j < n; ++j) {
+        for (int i = 0; i < m; ++i) {
+            for (int j = 0; j < n; ++j) {
                 if (matrix[i][j] == 0) {
                     row[i] = col[j] = true;
                 }
             }
         }
-        for (size_t i = 0; i < m; ++i) {
+        for (int i = 0; i < m; ++i) {
             if (row[i]) {
                 fill(&matrix[i][0], &matrix[i][0] + n, 0);
             }
         }
-        for (size_t j = 0; j < n; ++j) {
+        for (int j = 0; j < n; ++j) {
             if (col[j]) {
-                for (size_t i = 0; i < m; ++i) {
+                for (int i = 0; i < m; ++i) {
                     matrix[i][j] = 0;
                 }
             }
@@ -48,26 +48,26 @@ void print(vector<T> v) {
 }
 
 void setZeroes(vector<vector<int> > &matrix) {
-    const size_t m = matrix.size();
-    const size_t n = matrix[0].size();
+    const int m = matrix.size();
+    const int n = matrix[0].size();
     vector<bool> row(m, false);
     vector<bool> col(n, false);
 
-    for (size_t i = 0; i < m; ++i) {
-        for (size_t j = 0; j < n; ++j) {
+    for (int i = 0; i < m; ++i) {
+        for (int j = 0; j < n; ++j) {
             if (matrix[i][j] == 0) {
                 row[i] = col[j] = true;
             }
         }
     }
-    for (size_t i = 0; i < m; ++i) {
+    for (int i = 0; i < m; ++i) {
         if (row[i]) {
             fill(&matrix[i][0], &matrix[i][0] + n, 0);
         }
     }
-    for (size_t j = 0; j < n; ++j) {
+    for (int j = 0; j < n; ++j) {
         if (col[j]) {
-            for (size_t i = 0; i < m; ++i) {
+            for (int i = 0; i < m; ++i) {
                 matrix[i][j] = 0;
             }
         }
